Add new game elements Buff and Lightning
Buff is a new Sprite element, hide within a brick and appear
occasionally when the brick breaks.

There are two kinds of Buff, hide within different Bricks. They cannot
bounce back, and disappear when falling out of the stage. When the
Paddle touches a Buff, it becomes larger or carries "Lightning" for
some time, these two effects can be superimposed.

Lightning is a new Sprite element, both the main Ball and the Paddle
can carry it. When the main Ball bounced from the "Lightning" Paddle,
the Ball carries "Lightning", which makes it able to penetrate a few
Bricks.