Add audio and third level
Add effect audio and background audio to the game. Effect audio plays
when a button being clicked, a brick breaks or a buff is applied.
Background audio plays when the game starts.

Both types of audios can be muted in setting panel.

The third level has interesting layout and is also hard!